add 注解路由 和模型

+++ config/autoload/databases.php
@@ -13,9 +13,9 @@ declare(strict_types=1);
 return [
     'default' => [
         'driver' => env('DB_DRIVER', 'mysql'),
-        'host' => env('DB_HOST', 'localhost'),
+        'host' =>  env('DB_HOST', 'localhost'),
         'database' => env('DB_DATABASE', 'hyperf'),
-        'port' => env('DB_PORT', 3306),
+        'port' => (int) env('DB_PORT', 3306),
         'username' => env('DB_USERNAME', 'root'),
         'password' => env('DB_PASSWORD', ''),
         'charset' => env('DB_CHARSET', 'utf8'),
